Updated openstack/openstack Project: openstack/neutron  84b2f048e0561ab4e994e1379f08dd3795250a87
Wait for the watch process in test case

Because the _watch_process and the failing_process are asynchronous,
there might be a chance that failing_process exit and _watch_process
is not executed.

If the _watch_process is blocked, the method that will be asserted
will not be called. This will fail the UT, but it is intermittent.

Change-Id: Ic951c1b91c5a10462f548544a5e8d482c52ad665
Closes-Bug: #1519160