Updated openstack/openstack Project: openstack/nova  7bc2171869e2d7a670dc3c93494cf6ca05cac2fc
docs: add clarifications to move servers

This adds in some titles, to replace the bold text that simulated
heading. It also changes the text into use cases rather than actions.

It includes a some additions around:
* resource optimization use cases for live-migration
* data loss during evacuate
* downtime during resize
* clearer description of why a user would call shelve

This does not complete this doc, but should help improve the readability.

blueprint complete-todo-in-api-concept-doc

Change-Id: Ib2cf479c14a1c2f3b119ba7f549a00c9bd541cf9