Updated openstack/openstack Project: openstack/tripleo-heat-templates  a8bf039544ebf9b9e41a871445b6661182ad7044
Add net_vip_map_external to be used for an external balancer

Changes VipMap into a new NetVipMap resource which defaults to
being the same as the 'old' VipMap. An environment file can be
used to map NetVipMap instead to the net_vip_map_external.yaml
which allows for passing in explicit Virtual IP addresses.

It also ensures that references to the Virtual IPs are gathered
from the VipMap resource and allows for an empty ControlPlaneIP
parameter in the neutron port templates where it can be.

Co-Authored-By: Giulio Fidente <gfidente@redhat.com>

Change-Id: Ifad32e18f12b9997e3f89e4afe3ebc4c30e14a86