Updated openstack/openstack Project: openstack-infra/storyboard-webclient  1b786bbf7bf14b8ba275defa8d262a552ff1d42c
Align story subscription stars on right

Previously, when a user clicked on a project, the subscription stars
for the list of stories associated with it was on the left. This was
different to every other page that displayed subscription stars for
lists of items. This patch aligns the stars on the right.

Change-Id: I24f0dd67326d5348b3a6d73fb6a033f32adef770