Updated openstack/openstack Project: openstack/glance  467f240968f3b12b8a4d50ba2e48730794c0b81a
V1: Fix bad dates returning 500

Currently if you send a bad date format such as 'foo', as one of the
settable dates ie. `created_at`, `updated_at` and `deleted_at` it will
cause the server to return a 500.

This change means that the server will respond with a 400 and a helpful
message.

Change-Id: Ib342ef6640efbc3d76c4bafc622ccfdfa3d16ade
Closes-Bug: 1483649