Updated openstack/openstack Project: openstack/glance  d24cae18daeea161cc53e4e6bb083005af22751d
Add -constraints for CI jobs

This adds the constraints factor to the base section and duplicates
the sections required for CI with -constraints as tox does not
currently support factors in sections.  Work towards enabling factor
support in sections is currently stalled, as such we will need to
duplicate sections adding -constraints to facilitate running sections
with the constrained install_command.

Implements Blueprint: Requirements-Management

Change-Id: Ibbeabdf65c1405bf4d35e21375636dcc4583bbc2