Updated openstack/openstack Project: openstack-infra/project-config  aa01c1d15e1a2b8a4a02801c340cbfda36b0616f
[Rally] add new job for verification

Currently we have only one job to check that `rally verify` command works
well. It launches small number of tempest tests, which is not enought to
check changes in generation tempest config, so this patch adds new job
which will launch full tempest run when code related to verification is
changed in rally.

Change-Id: Idb0e6f955e23055a2d98295d2628a30973ecb115