Updated openstack/openstack Project: openstack/oslo.policy  0e9f395f3c20d062ddff42e9b4b4cce85a20216a
Clarify usage docs

The policy docs were confusing because they start describing the
policy rule expressions but reference only one format. Rather than
provide a short and incorrect description in the intro, reference
the section that describes the format in full.

Also, the section that describes the rule format is moved right
after the intro since this is what the reader is going to be
looking for.

Also, the "Policy Rule Expressions" section is clarified. Since the
string representation is preferred, describe that first.

Change-Id: I69b74e2d0bd9cbe8fe447a1e0f8fa5ce2470b431