Updated openstack/openstack Project: openstack/ceilometer  37905010bcafe822d53d249ab80f583c474d14f2
devstack config for dogpile cache

Cause devstack to configuration a [cache] section in ceilometer.conf
if CEILOMETER_CACHE_BACKEND is set. Currently set to default to
using dogpile.cache.redis and a redis server on localhost.

This options are not sufficient for generic use as the
"backend_argument"s are not controlled. For general use it is
probably best that tools just modify their local.conf to post config
ceilometer.conf.

Change-Id: I058e60f9440350f80262039171716a3ceb4dec62