Updated openstack/openstack Project: openstack-dev/devstack  6bfd30b7adbb12c6a7e917d386ee039eeaff1c41
Consolidate two /sbin/ip commands into a single one

When determining the primary IP/IPv6 addresses on a system,
we can use the /sbin/ip command to filter them for us.  The
resulting address is parsed the same way for both address
families, so we can use just a single command.

Change-Id: I0471ff5a258a16a23061941ac38063dbf3d7c233