Updated openstack/openstack Project: openstack/openstack-manuals  9de301547fbb080533dab092fc6e81348f781e99
Refactoring HP 3PAR Manila Driver to be HPE

The HP 3PAR Manila driver was recently refactored to be HPE.

The same refactoring is being done to the documentation with
this patch. It is being done because the company managing the
HP 3PAR Manila driver is now Hewlett Packard Enterprise (HPE).

Closes-Bug: #1517283
Change-Id: I0253a5fba8218960014d0b06cbc92baa6a775a2c