Updated openstack/openstack Project: openstack/heat  22ec82ca25f78eae94a57fb02b97b0c8793912dd
Convergence: Re-trigger failed resource for latest traversal

Presently, when a resource of previous traversal completes its action
successfully we re-trigger this resource for latest traversal.(since
the latest traversal will be waiting for its completion)

However, if a resource of previous traversal fails we do not
re-trigger which leads to latest traversal waiting endlessly.

This patch re-triggers the resource for latest traversal even when
the resource fails.

Change-Id: I9f70878ad7f1ff7c2facb950e496681425b54fc4
Partial-Bug: #1512343