Updated openstack/openstack Project: openstack/python-glanceclient  101528ee6309b4011a007150c93c3396036bf1b0
Ensure that identity token in header is not an unicode string

We need all the headers to be safe strings so they can be joined
together and not become an unicode string in doing so.

This fixes a bug when creating an image with non-ascii characters in the
name.

This is required for python 2.6 compatibility.

Change-Id: I66ebc27edf4ccd8f903399da58705711c372536d
Closes-Bug: 1448080