Updated openstack/openstack Project: openstack/nova  92d2de680dddb9305cd6fe9b422d97f660af8a7b
Detach volume after deleting instance with no host

If an instance is booted from a volume, shelved, and goes into an error
state due to some reason. Volume from which instance is booted, remains
in-use state even the instance is deleted because instance has no host
associated with it.

Called _local_delete() to detach volume and destroy bdm if instance is
in shelved_offloaded state or has no host associated with it. This will
cleanup both volumes and the networks.

Note:
I had submitted same patch [1] earlier which was reverted [2] due to a
race condition on jenkins if an instance is deleted when it is in
building state. In this patch I have fixed the failure of race condition
by reverting the ObjectActionError exception handling in _delete.

[1] Ic630ae7d026a9697afec46ac9ea40aea0f5b5ffb
[2] Id4e405e7579530ed1c1f22ccc972d45b6d185f41

Closes-Bug: 1404867
Closes-Bug: 1408527
Closes-Bug: 1458308
Change-Id: Ic107d8edc7ee7a4ebb04eac58ef0cdbf506d6173