Updated openstack/openstack Project: openstack/openstack-manuals  3b89e57056da413e90e1270c241540bb9e9544f4
[install] Updates the keystone install

This patch updates the commands made by the postinst of Keystone,
as the package moved to use the Keystone v3 API.

Note that still missing is the use of Apache instead of the
eventlet setup for Keystone. Please bare with me, and accept
this as an incremental step, as this will come later on.

Change-Id: I497f671b4f96d700be189d8e363a8c75c513e207