Updated openstack/openstack Project: openstack-infra/shade  2c753699ac9c3c4209d5f070855f4315ea468efa
Don't double-print exception subjects

We're creating an error message string that contains the shade message
and the underlying message AND we're passing the underlying message in
so that we append it as part of the Inner Exception string. So we wind
up with something like:

  Error in creating volume: Invalid input received: 'size' parameter
  must be between 75 and 1024 (HTTP 400) (Request-ID:
  req-ae3f3489-997b-4038-8c9f-efbe7dd73a70) (Inner Exception: Invalid
  input received: 'size' parameter must be between 75 and 1024)

Which is a bit insane. With this change, the above becomes:

  Error in creating volume (Inner Exception: Invalid input received:
  'size' parameter must be between 75 and 1024 (HTTP 400) (Request-ID:
  req-d4da66b2-41eb-44c8-85f1-064454af5a1c))

And if the shade_exceptions context manager is invoked with no message,
the output will be:

  Invalid input received: 'size' parameter must be between 75 and 1024
  (HTTP 400) (Request-ID: req-d4da66b2-41eb-44c8-85f1-064454af5a1c))

Change-Id: I6bd70b70585722d2266ef08496b6860aeeab1824