Updated openstack/openstack Project: openstack-infra/system-config  f1b9b864f7ac5c4886be685eb82ff12ef0f3d734
Translate the rest of run_all.sh to ansible

There are a few things that are run as part of run_all.sh that are
not logged into puppet_run_all.log - namely git cloning, module installation
and ansible role installation. Let's go ahead and do those in a playbook
so that we can see their output while we're watching the log file.

Change-Id: I6982452f1e572b7bc5a7b7d167c1ccc159c94e66