Updated openstack/openstack Project: openstack-infra/shade  07ac77ca765df2f9cf46b0210d40ef2f96fe5586
Add functional tests for boot from volume

Test both that we can boot from volume from a pre-existing image
with and without auto-termination. Also that manually making the volume
from the image and then booting from it also works.

Also, add a cache invalidation that testing showed to be an issue.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>

Change-Id: Ia0f5801c17edacf8cce67c7594657e7ae41e0f18