Updated openstack/openstack Project: openstack-infra/subunit2sql  5b9f790b1ddb716d98e6a19c65de13112661c35c
Return Run.uuid in public methods

In the UUID->BIGINT migration, we kept Run.uuid because it is intended
as a stable user-accessible handle for that particular run. We failed
to catch all of the places that leaked the table ID column, so there
are a few places now that are getting ints which can't subsequently
be looked up properly with sql2subunit.

This also fixes some minor issues with the get_run_id_from_uuid()
method that were making the behavior of the function deviate from the
documented and expected behavior.

Change-Id: I19162bf2f2677d86eb7daca49aab6c3cb23fc206