Updated openstack/openstack Project: openstack-infra/project-config  c461411578375068ad0d7c7f6c63830139ec6432
Make Heat Convergence gate voting

The gate seems to be stable now with all tests passing.
Make the convergence gate voting.

Also disable for liberty release, because during
L release this feature was in development and the job
was a additional source for work feedback.

We can not port all necessary changes to 
liberty branch to make it work, so best to skip it

Change-Id: I2845e9d335eb5fb0a627c4097048e21d2b7399fe
Depends-On: Ia973b3033b031eff7b3713c347630bf5d91d0377