Updated openstack/openstack Project: openstack/ironic  63a16d2c6b4b67d0255e9b8d65b2abc38bee9a0d
Add a test to enforce object version bump correctly

Add a test to help developers to check if the change of objects
need a version bump.

Any changes of object fields or remotable methods need to bump
its version. Use oslo_versionobjects to generate the hashes and
compare with a pre-define hash map.

Change-Id: I5c12ab72937f05ba46eb58d049f9b7fc7c806218
Closes-Bug: 1502895
Implements: blueprint online-upgrade-support