Updated openstack/openstack Project: openstack/python-ceilometerclient  cdb47d9646236e19208b7e49ffc6c7d42000dd89
Token will be lost when constructing a ceilometer client

If parameters os_token and os_endpoint were passed to
ceilometerclient.client.get_client to construct a ceilometer client, the
token will be wrongly replaced with None in method
ceilometerclient.client._adjust_kwargs, which will cause the
authentication error when use the instance of ceilometer client later.
This patch fixed this problem.

Change-Id: Ib15cdd73b62b7f0b86d0bd6f19d344fd24a665f6
Closes-Bug: #1519729