Updated openstack/openstack Project: openstack/keystonemiddleware  0981dedf15f9bcc07e7d6554ff23faad4fd38af5
Add a mock-fixture for keystonemiddleware auth_protocol

This adds a test-usable fixture that will patch the base protocol
to simply return the token data specified rather than reaching out
to the keystone server (or lack of keystone server). This is intended
to allow projects such as ceilometer to function without needing to
patch the internals of keystonemiddleware (memcache interface).

Change-Id: I1859ba9e3941eb5400f4cece5f88a5e747836fea