Updated openstack/openstack Project: openstack-infra/system-config  0fbbcb0ea983d95579093d43d3deaa6917505209
Add shade and openstack inventory to system

We're not ready to move from puppet inventory to openstack inventory
just yet, so don't actually swap the dynamic inventory plugin. But, add
it to the system so that running manual tests of all of the pieces is
possible.

Add the currently administratively disabled hosts to the disabled group
so that we can verify this works.

Change-Id: I73931332b2917b71a008f9213365f7594f69c41e