Updated openstack/openstack Project: openstack/horizon  746cc68c562f990c2c0c1afab12f5ec6a419dec1
Change Permanent Cookie Contain Sensitive Info

Right now, the 'csrftoken' cookie is stored on disk as a permanent
cookie. There is a risk for sensitive session information (cookies)
that are persisted on disk as permanent cookies.

This fixes this issue by storing the cookies in-memory instead of
in persistent storage.

Change-Id: Ia45b09571d495d4f98b60545903af72eb0f061c2
Closes-Bug: #1369865