Updated openstack/openstack Project: openstack/keystone  0eead474c0e5f8bb2509b4557dadea4075038243
Needn't care about the sequence for cache validation

The testcases were added to check cache invalidation [1] but fail
occasionally with the assertion where ordered sequence specific
comparison is required, this is because the value fetched from DB
or cache not always match with the sequence of when the value was
created.

Update the testcases to use unordered sequence specific comparison.

[1] https://review.openstack.org/#/c/245633/

Change-Id: Iffa3ccaa1b0a6627c34dc8902c507ce741442cc0
Closes-Bug: #1520850