Updated openstack/openstack Project: openstack/cinder  98864f49aa53f86fecda86e4048dc60a692d094a
Declare multiattach is True in Dell Eqlx driver

An Eqlx volume can be attached to multi instances, however,
this feature is not declared in eqlx driver. As a result,
if a user create a multi-attached volume using eqlx as
backend, they will be failed.

In this commit, we declared multiattach is True in eqlx
driver.

Change-Id: Ia8e94e739c2573f1f26d8cb2713a6b851adf9428
Closes-Bug: #1519570