Updated openstack/openstack Project: openstack-infra/project-config  832d7de8fb016e757124db9fdc7af2eb6e8cbc4a
Update to Fedora 23

The prior change (Ifa6367488165df6cad18725a9191b426a41c08bf) has
disabled F22 nodes.

This adds F23 nodes (just one for inital testing) and modifies the
fedora-22 based jobs to use the new nodes.

(note the depends-on, required for building f23 fedora-minimal, is
available in dib 1.4.0, so this is ready to go)

Depends-On: I1a6e45d04755515286b3d49f8280c16b527e2f48
Change-Id: I3de17c2d670d8cb0f28ffb5e1541751bc7f4d73e