Updated openstack/openstack Project: openstack/nova-specs  8b3c2633b7d7d4a17186ac565fb0e647a7fffb9e
Parallel Scheduler for cells v2 backlog spec

To help share ideas, adding a backlog spec to document the problems
with the current filter scheduler.

It also documents a number of possible solutions that have been
discussed in recent years, that fit together to fix the issues for
public clouds that currently use cells v1 and are looking at moving to
cells v2 in the M release.

Please note: there is currently no blueprint associated with this
specification, as its still in the problem definition and idea gathering
phase.

Change-Id: Ie52b862ace7801c17e3844d5c9ed8cb25a0592f1