Updated openstack/openstack Project: openstack/cinder-specs  be39b6c16a36c9b9a4fa5cb3fce9737245507d11
cinder-api-microversions

Many changes to the Cinder REST API require changes to the consumers of the API.
For example, If we need to add a required parameter to a method that is called
by Nova, we'd need both the Nova calling code and the cinderclient that
Nova uses to change. But newer Cinder versions with the change must work with
older Nova versions, and there is no mechanism for this at the moment. Adding
microversions will solve this problem.

With microversions, the highest supported version will be negotiated by a field
in the HTTP header that is sent to the Cinder API. In the case where the field
'versions' is not sent (i.e. clients and scripts that pre-date this change),
then the lowest supported version would be used. This means that our current
Cinder API v2 would be the default, and consumers of the API that wished to
use a newer version could do so.

Implements: https://blueprints.launchpad.net/cinder/+spec/cinder-api-microversions
Change-Id: I2eb6b3aa86f8f6224dd223a8b8cef34ea15ce1e2