Updated openstack/openstack Project: openstack/ironic  215f37ba043e63cd520c65caaa163dda3f583f82
Fall back to old boot.ipxe behaviour if inc command is not found

We introduced looping over all NIC's as on older iPXE ROM's ${mac} variable
does not always point to the PXE-booting NIC. However, CentOS/RHEL ship
even older iPXE ROM that does not even have inc command at all.

This patch restores the old behaviour of relying on ${mac} variable when
inc command cannot be executed. It's still not perfect and might break
sometimes, but it won't be completely broken. Nothing is changes for users
for which inc command is already working.

Change-Id: I30b8c0e4ab076684ea9baa042e5a9a4f44da9a3a
Closes-Bug: #1507738