Updated openstack/openstack Project: openstack/pycadf  b6a74b729c0d00fd9374046832ee7cef3a9a0528
make generate_uuid return valid uuid

original design prepended a namespace to the beginning of uuid.
this scoped the uuid to a namespace but it is technically invalid.

this patch drops the behaviour of prepending string. rather, it
takes the namespace, hashes it, and uses uuid v5 to generate valid
uuid based on that namespace

Change-Id: Ibde222c91522fa3c4a1720b6a8da81dac31560b9
Closes-Bug: #1504889