Updated openstack/openstack Project: openstack/sahara-dashboard  0d3bead819590a0dc9ebbafe26fae5b8dcaba278
Moving exception specification

Horizon defines classes of exception from the services it
communicates with. In the plugin case, the plugin needs to set
up including plugin content specific exceptions to the classes of
exceptions in Horizon. This patch includes this content.

Setting up an extensible exception list in exceptions.py and
including the exceptions in the configuration files.

Change-Id: Ie91cfe1e0f34a710a96b4a43cf62af35d989920c