Updated openstack/openstack Project: openstack/trove  f66701218bf0e8fbd06305115b37c2ba7e499dd2
Change reporting of unhandled logging

Unhandled logging messages generated from sources other than Trove's
tests were flooding the log with numerous messages that included
complete backtraces. These backtraces are over 150mb when you run tox
-e py27. Unfortunately, while they are an indication of a problem,
they are most often not things that can be fixed in the Trove project.

To fix this, at least in the short term, this change only reports
anything if the logged message is coming from the context of a test.

Also, this change disables back traces in the messages being reported
unless an optional enables_backtrace is set. By default it set to
False and can only be enabled with a code change (i.e. no
configuration setting).

Change-Id: I9452f03bdd5365991e3c4831aa4936de3713612e
Closes-Bug: 1521373